requirements.txt
# Voice Server dependencies
flask
flask-cors
torch
openai-whisper
gTTS

# llama-cpp-python

# CMAKE_ARGS="-DLLAMA_METAL=on" pip install llama-cpp-python --no-cache-dir

Pillow
requests
numpy
resampy
soundfile

# Voice Client dependencies
sounddevice
pygame
flask-sock 
websocket-client
pyttsx3 
sounddevice
onnxruntime 
transformers
colorama
jiwer
silero-vad
noisereduce
pynput
mlx-audio
opencv-python
playwright 
schedule 
# playwright install chromium
browser-cookie3
pyautogui 
pyperclip
langgraph 
langchain-openai
pytesseract